Move check coords to shader_lib
